Do not rely on Sys.BINDIR for docs paths
In practice, the Julia binary might not live in the same directory as
the Julia documentation source files. E.g. when you are building the
documentation from a tarball with Julia installed via the package
manager.

Changing it to @__DIR__ makes sure that the path is always relative to
the documentation Markdown files. The downside is that currently
@__DIR__/@__FILE__ in Documenter at-blocks give a paths in the output
directory and not in the source directory.

Fixes #29694.